requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bqplot==0.12.43
branca==0.7.2
build==1.2.2.post1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==6.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
durationpy==0.9
earthengine-api==0.1.405
eerepr==0.0.4
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.112.1
fastapi-cli==0.0.5
fastkml==0.12
feedparser==6.0.11
filelock==3.16.1
fiona==1.9.6
Flask==3.0.3
flask-restx==1.3.0
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gnews==0.3.7
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-genai==1.11.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.22.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.18.2
ipython==8.25.0
ipytree==0.2.2
ipywidgets==8.1.3
itsdangerous==2.2.0
jaraco.context==6.0.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.2.16
langchain-chroma==0.1.4
langchain-community==0.2.10
langchain-core==0.2.43
langchain-experimental==0.0.63
langchain-google-genai==1.0.3
langchain-groq==0.3.2
langchain-mcp-adapters==0.0.9
langchain-openai==0.1.17
langchain-text-splitters==0.2.4
langgraph==0.2.22
langgraph-checkpoint==1.0.10
langsmith==0.1.137
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mesop==0.9.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.5.5
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.37.1
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
ordered-set==4.1.0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
plotly==5.22.0
posthog==3.7.0
prompt_toolkit==3.0.45
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
pydeck==0.9.1
pygeoif==0.7
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==5.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyshp==2.3.1
PySocks==1.7.1
python-box==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
selenium==4.23.1
sentence-transformers==3.2.1
sgmllib3k==1.0.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.38.0
sympy==1.13.1
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
torch==2.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
traittypes==0.2.1
transformers==4.45.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
watchdog==4.0.1
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wolframalpha==5.1.3
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.14.2
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.20.2
